// types/storybrand.ts
// Character type for storybrand
export interface Character {
    character: string;
    want: string;
    externalProblem: string;
    internalProblem: string;
    philosophicalProblem: string;
    empathyStatement: string;
    authorityStatement: string;
    plan: string;
    callToActions: string;
    avoidFailure: string;
    successLooksLike: string;
    oneLiner: string;
    elevatorPitch: string;
    landingPage: string;
  }